requirements.txt
python-telegram-bot
python-dotenv
fastapi[standard]
ollama
groq
supabase